feat: 한개의 모달 제어를 위한 useModal 구현
